#134 fallback to a different "Scripts" directory if default one is not writable
